revise layers doc
I have in mind something like a message storage layer or module -- I'm
not sure how this fits in, but it seems like messages should always
include the hash(es) of their parent(s), so that the message history
is a directed acyclic graph (DAG).

I also have in mind something like a tree or hypergraph structure for
storing timelines or worldlines of events or commands.  I'm also not
sure how this fits in.

Discuss how these layers or modules might interact, and pros and cons
of each approach.